code/backend/requirements.txt
fastapi==0.104.1
uvicorn>=0.36.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
discord.py==2.3.2
httpx==0.25.1
websockets==12.0
openai>=1.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
weasyprint>=60.0
pillow>=10.0.0
uvicorn-worker==0.4.0
